refactor: folder structure routing
